Module Async.Ppx_log_syntax

include Ppx_log_types.S with type t = Async_unix.Log.t and type return_type = unit and type Global.return_type = unit

type t = Async_unix.Log.t

type return_type = unit

val would_log : t -> Ppx_log_types.level Base.option -> Base.bool

val default : return_type

val sexp : ?level:Ppx_log_types.level -> ?pos:Base.Source_code_position.t -> t -> Sexplib.Sexp.t -> return_type

module Global : sig ... end

module No_global : sig ... end

If you wish you prevent global logging with ppx_log, you can open this module and any use of global logging will return a warning type `Do_not_use_because_it_will_not_log (and will not actually log anything).