Module Async.Sys

This module overrides everything in the Core.Sys module that might block. Functions do the same thing as their counterparts in Core.Sys, but instead return deferreds. For a description of their semantics see the documentation for the Core.Sys module.

val argv : string array

val get_argv : unit -> string array

val executable_name : string

val file_exists : ?⁠follow_symlinks:bool -> string -> [ `Yes | `No | `Unknown ] Async_kernel.Deferred.t

val file_exists_exn : ?⁠follow_symlinks:bool -> string -> bool Async_kernel.Deferred.t

val when_file_exists : ?⁠follow_symlinks:bool -> ?⁠poll_delay:Core.Time.Span.t -> string -> unit Async_kernel.Deferred.t

when_file_exists ?poll_delay file returns a deferred that becomes determined when file exists. The default poll delay is 0.5 seconds. It raises an exception if it can not check whether the file is there, in the same cases file_exists returns `Unknown.

val when_file_changes : ?⁠poll_delay:Core.Time.Span.t -> string -> Core.Time.t Async_kernel.Pipe.Reader.t

when_file_changes file polls file using stat and writes file's mtime to the pipe every time it changes. The first time in the pipe will be file's current mtime. To stop polling, close the pipe.

val is_directory : ?⁠follow_symlinks:bool -> string -> [ `Yes | `No | `Unknown ] Async_kernel.Deferred.t

val is_directory_exn : ?⁠follow_symlinks:bool -> string -> bool Async_kernel.Deferred.t

val is_file : ?⁠follow_symlinks:bool -> string -> [ `Yes | `No | `Unknown ] Async_kernel.Deferred.t

val is_file_exn : ?⁠follow_symlinks:bool -> string -> bool Async_kernel.Deferred.t

val remove : string -> unit Async_kernel.Deferred.t

val rename : string -> string -> unit Async_kernel.Deferred.t

val getenv : string -> string option

val getenv_exn : string -> string

val command : string -> int Async_kernel.Deferred.t

val command_exn : string -> unit Async_kernel.Deferred.t

val quote : string -> string

val chdir : string -> unit Async_kernel.Deferred.t

val getcwd : unit -> string Async_kernel.Deferred.t

val readdir : string -> string array Async_kernel.Deferred.t

val ls_dir : string -> string list Async_kernel.Deferred.t

val home_directory : unit -> string Async_kernel.Deferred.t

val opaque_identity : 'a -> 'a

val interactive : bool Core.ref

val os_type : string

val word_size : int

val int_size : int

val big_endian : bool

val ocaml_version : string

val execution_mode : unit -> [ `Bytecode | `Native ]

val c_int_size : unit -> int