Module Async_command.Spec

The old interface for command-line specifications -- Do Not Use.

This interface should not be used. See the Param module for the new way to do things.

Command parameters

type 'a param = 'a Param.t

Specification of an individual parameter to the command's main function.

include Param.S with type 'a t := 'a Param.t

Command.Param is intended to be used with the [%map_open] syntax defined in ppx_let, like so:

include Base.Applicative.S with type 'a t := 'a Param.t

val return : 'a -> 'a Param.t

val map : 'a Param.t -> f:('a -> 'b) -> 'a Param.t

val both : 'a Param.t -> 'a Param.t -> 'a Param.t

val (<*>) : 'a Param.t -> 'a Param.t -> 'a Param.t

same as apply

val (<*) : 'a Param.t -> 'a Param.t -> 'a Param.t

val (*>) : 'a Param.t -> 'a Param.t -> 'a Param.t

val (>>|) : 'a Param.t -> ('a -> 'b) -> 'a Param.t

val apply : 'a Param.t -> 'a Param.t -> 'a Param.t

val map2 : 'a Param.t -> 'a Param.t -> f:('a -> 'b -> 'c) -> 'a Param.t

val map3 : 'a Param.t -> 'a Param.t -> 'a Param.t -> f:('a -> 'b -> 'c -> 'd) -> 'a Param.t

val all : 'a Param.t list -> 'a Param.t

val all_unit : 'a Param.t list -> 'a Param.t

module Applicative_infix : sig ... end

Various internal values

val help : 'a Param.t

The help text for the command.

val path : 'a Param.t

The subcommand path of the command.

val args : 'a Param.t

The arguments passed to the command.

val flag : ?⁠aliases:Base.String.t Base.List.t -> ?⁠full_flag_required:Base.Unit.t -> Base.String.t -> 'a Flag.t -> doc:Base.String.t -> 'a Param.t

flag name spec ~doc specifies a command that, among other things, takes a flag named name on its command line. doc indicates the meaning of the flag.

All flags must have a dash at the beginning of the name. If name is not prefixed by "-", it will be normalized to "-" ^ name.

Unless full_flag_required is used, one doesn't have to pass name exactly on the command line, but only an unambiguous prefix of name (i.e., a prefix which is not a prefix of any other flag's name).

NOTE: the doc for a flag which takes an argument should be of the form arg_name ^ " " ^ description where arg_name describes the argument and description describes the meaning of the flag.

NOTE: flag names (including aliases) containing underscores will be rejected. Use dashes instead.

NOTE: "-" by itself is an invalid flag name and will be rejected.

val flag_optional_with_default_doc : ?⁠aliases:Base.String.t Base.List.t -> ?⁠full_flag_required:Base.Unit.t -> Base.String.t -> 'a Arg_type.t -> ('a -> Core_kernel.Sexp.t) -> default:'a -> doc:Base.String.t -> 'a Param.t

flag_optional_with_default_doc name arg_type sexp_of_default ~default ~doc is a shortcut for flag, where:

The Flag.t is optional_with_default default arg_type

The doc is passed through with an explanation of what the default value appended.

val anon : 'a Anons.t -> 'a Param.t

anon spec specifies a command that, among other things, takes the anonymous arguments specified by spec.

module If_nothing_chosen : sig ... end

val choose_one : 'a Param.t Base.List.t -> if_nothing_chosen:('a, 'b) If_nothing_chosen.t -> 'a Param.t

choose_one clauses ~if_nothing_chosen expresses a sum type. It raises if more than one of clauses is Some _. When if_nothing_chosen = `Raise, it also raises if none of clauses is Some _.

val and_arg_names : 'a Param.t -> 'a Param.t

and_arg_names t returns both the value of t and the names of the arguments that went into t. Useful for errors that reference multiple params.

val and_arg_name : 'a Param.t -> 'a Param.t

Like and_arg_names, but asserts that there is exactly one name.

val const : 'a -> 'a Param.t

Superceded by return, preserved for backwards compatibility.

val pair : 'a Param.t -> 'b Param.t -> ('a * 'b) Param.t

Superceded by both, preserved for backwards compatibility.

Command specifications

type (-'main_in, +'main_out) t

Composable command-line specifications.

Ultimately one forms a basic command by combining a spec of type ('main, unit -> unit) t with a main function of type 'main; see the basic function below. Combinators in this library incrementally build up the type of main according to what command-line parameters it expects, so the resulting type of main is something like:

arg1 -> ... -> argN -> unit -> unit

It may help to think of ('a, 'b) t as a function space 'a -> 'b embellished with information about:

how to parse the command line

what the command does and how to call it

how to autocomplete a partial command line

One can view a value of type ('main_in, 'main_out) t as a function that transforms a main function from type 'main_in to 'main_out, typically by supplying some arguments. E.g., a value of type Spec.t might have type:

(arg1 -> ... -> argN -> 'r, 'r) Spec.t

Such a value can transform a main function of type arg1 -> ... -> argN -> 'r by supplying it argument values of type arg1, ..., argn, leaving a main function whose type is 'r. In the end, Command.basic takes a completed spec where 'r = unit -> unit, and hence whose type looks like:

(arg1 -> ... -> argN -> unit -> unit, unit -> unit) Spec.t

A value of this type can fully apply a main function of type arg1 -> ... -> argN -> unit -> unit to all its arguments.

The final unit argument allows the implementation to distinguish between the phases of (1) parsing the command line and (2) running the body of the command. Exceptions raised in phase (1) lead to a help message being displayed alongside the exception. Exceptions raised in phase (2) are displayed without any command line help.

The view of ('main_in, main_out) Spec.t as a function from 'main_in to 'main_out is directly reflected by the step function, whose type is:

val step : ('m1 -> 'm2) -> ('m1, 'm2) t

val empty : ('m, 'm) t

The empty command-line spec.

val (++) : ('m1, 'm2) t -> ('m2, 'm3) t -> ('m1, 'm3) t

Command-line spec composition.

val (+>) : ('m1, 'a -> 'm2) t -> 'a Param.t -> ('m1, 'm2) t

Adds a rightmost parameter onto the type of main.

val (+<) : ('m1, 'm2) t -> 'a Param.t -> ('a -> 'm1, 'm2) t

Adds a leftmost parameter onto the type of main.

This function should only be used as a workaround in situations where the order of composition is at odds with the order of anonymous arguments because you're factoring out some common spec.

val step : ('m1 -> 'm2) -> ('m1, 'm2) t

Combinator for patching up how parameters are obtained or presented.

Here are a couple examples of some of its many uses:

introducing labeled arguments

step (fun m v -> m ~foo:v)
               +> flag "-foo" no_arg : (foo:bool -> 'm, 'm) t

prompting for missing values

step (fun m user -> match user with
                 | Some user -> m user
                 | None -> print_string "enter username: "; m (read_line ()))
               +> flag "-user" (optional string) ~doc:"USER to frobnicate"
               : (string -> 'm, 'm) t

A use of step might look something like:

step (fun main -> let ... in main x1 ... xN) : (arg1 -> ... -> argN -> 'r, 'r) t

Thus, step allows one to write arbitrary code to decide how to transform a main function. As a simple example:

step (fun main -> main 13.) : (float -> 'r, 'r) t

This spec is identical to const 13.; it transforms a main function by supplying it with a single float argument, 13.. As another example:

step (fun m v -> m ~foo:v) : (foo:'foo -> 'r, 'foo -> 'r) t

This spec transforms a main function that requires a labeled argument into a main function that requires the argument unlabeled, making it easily composable with other spec combinators.

val wrap : (run:('m1 -> 'r1) -> main:'m2 -> 'r2) -> ('m1, 'r1) t -> ('m2, 'r2) t

Combinator for defining a class of commands with common behavior.

Here are two examples of command classes defined using wrap:

print top-level exceptions to stderr

wrap (fun ~run ~main ->
                 Exn.handle_uncaught ~exit:true (fun () -> run main)
               ) : ('m, unit) t -> ('m, unit) t

iterate over lines from stdin

wrap (fun ~run ~main ->
                 In_channel.iter_lines stdin ~f:(fun line -> run (main line))
               ) : ('m, unit) t -> (string -> 'm, unit) t

module Arg_type : module type of Arg_type with type 'a t = 'a Arg_type.t

include module type of Arg_type.Export

val string : Base.String.t Arg_type.t

val int : Base.Int.t Arg_type.t

Beware that an anonymous argument of type int cannot be specified as negative, as it is ambiguous whether -1 is a negative number or a flag. (The same applies to float, time_span, etc.) You can use the special built-in "-anon" flag to force a string starting with a hyphen to be interpreted as an anonymous argument rather than as a flag, or you can just make it a parameter to a flag to avoid the issue.

val char : Base.Char.t Arg_type.t

val float : Base.Float.t Arg_type.t

val bool : Base.Bool.t Arg_type.t

val date : Core_kernel.Date.t Arg_type.t

val percent : Core_kernel.Percent.t Arg_type.t

val host_and_port : Core_kernel.Host_and_port.t Arg_type.t

val sexp : Core_kernel.Sexp.t Arg_type.t

val sexp_conv : (Core_kernel.Sexp.t -> 'a) -> 'a Arg_type.t

type 'a flag = 'a Flag.t

A flag specification.

include module type of Flag with type 'a t := 'a flag

val required : 'a Arg_type.t -> 'a flag

Required flags must be passed exactly once.

val optional : 'a Arg_type.t -> 'a flag

Optional flags may be passed at most once.

val optional_with_default : 'a -> 'a Arg_type.t -> 'a flag

optional_with_default flags may be passed at most once, and default to a given value.

val listed : 'a Arg_type.t -> 'a flag

listed flags may be passed zero or more times.

val one_or_more : 'a Arg_type.t -> 'a flag

one_or_more flags must be passed one or more times.

val no_arg : 'a flag

no_arg flags may be passed at most once. The boolean returned is true iff the flag is passed on the command line.

val no_arg_register : key:'a Core_kernel.Univ_map.With_default.Key.t -> value:'a -> 'a flag

no_arg_register ~key ~value is like no_arg, but associates value with key in the autocomplete environment.

val no_arg_some : 'a -> 'a flag

no_arg_some value is like no_arg, but will return Some value if the flag is passed on the command line, and return None otherwise.

val no_arg_abort : exit:(Base.Unit.t -> Core_kernel.Nothing.t) -> 'a flag

no_arg_abort ~exit is like no_arg, but aborts command-line parsing by calling exit. This flag type is useful for "help"-style flags that just print something and exit.

val escape : 'a flag

escape flags may be passed at most once. They cause the command line parser to abort and pass through all remaining command line arguments as the value of the flag.

A standard choice of flag name to use with escape is "--".

val map_flag : 'a flag -> f:('a -> 'b) -> 'a flag

map_flag flag ~f transforms the parsed result of flag by applying f.

val flags_of_args_exn : Core_kernel.Arg.t Base.List.t -> ('a, 'a) t

flags_of_args_exn args creates a spec from Caml.Arg.ts, for compatibility with OCaml's base libraries. Fails if it encounters an arg that cannot be converted.

NOTE: There is a difference in side effect ordering between Caml.Arg and Command. In the Arg module, flag handling functions embedded in Caml.Arg.t values will be run in the order that flags are passed on the command line. In the Command module, using flags_of_args_exn flags, they are evaluated in the order that the Caml.Arg.t values appear in args.

type 'a anons = 'a Anons.t

A specification of some number of anonymous arguments.

include module type of Anons with type 'a t := 'a anons

val (%:) : Base.String.t -> 'a Arg_type.t -> 'a anons

(name %: typ) specifies a required anonymous argument of type typ.

The name must not be surrounded by whitespace; if it is, an exn will be raised.

If the name is surrounded by a special character pair (<>, {}, [] or (),) name will remain as-is, otherwise, name will be uppercased.

In the situation where name is only prefixed or only suffixed by one of the special character pairs, or different pairs are used (e.g., "<ARG]"), an exn will be raised.

The (possibly transformed) name is mentioned in the generated help for the command.

val sequence : 'a anons -> 'a anons

sequence anons specifies a sequence of anonymous arguments. An exception will be raised if anons matches anything other than a fixed number of anonymous arguments.

val non_empty_sequence_as_pair : 'a anons -> 'a anons

non_empty_sequence_as_pair anons and non_empty_sequence_as_list anons are like sequence anons except that an exception will be raised if there is not at least one anonymous argument given.

val non_empty_sequence_as_list : 'a anons -> 'a anons

val maybe : 'a anons -> 'a anons

(maybe anons) indicates that some anonymous arguments are optional.

val maybe_with_default : 'a -> 'a anons -> 'a anons

(maybe_with_default default anons) indicates an optional anonymous argument with a default value.

val t2 : 'a anons -> 'a anons -> 'a anons

val t3 : 'a anons -> 'a anons -> 'a anons -> 'a anons

val t4 : 'a anons -> 'a anons -> 'a anons -> 'a anons -> 'a anons

val map_anons : 'a anons -> f:('a -> 'b) -> 'a anons

map_anons anons ~f transforms the parsed result of anons by applying f.

val to_param : ('a, 'r) t -> 'a -> 'r Param.t

val of_param : 'r Param.t -> ('r -> 'm, 'm) t