Parameter Configure.1-Config

val default_seed : Core_kernel.Quickcheck_intf.seed

default_seed is used initialize the pseudo-random generator that chooses random values from generators, in each test that is not provided its own seed.

val default_sizes : Base.Int.t Core_kernel.Sequence.t

default_sizes determines the default sequence of sizes used in generating values.

val default_trial_count : Base.Int.t

default_trial_count determines the number of trials per test, except in tests that explicitly override it.

val default_can_generate_trial_count : Base.Int.t

default_can_generate_trial_count determines the number of trials used in attempts to generate satisfying values, except in tests that explicitly override it.

val default_shrink_attempts : Core_kernel.Quickcheck_intf.shrink_attempts

default_shrink_attempts determines the number of attempts at shrinking when running test or iter with ~shrinker and without ~shrink_attempts