Module Log_extended.Command

type console_style =

| Plain

| Color

val sexp_of_console_style : console_style -> Sexplib0.Sexp.t

val console_style_of_sexp : Sexplib0.Sexp.t -> console_style

type console_output =

| No

| Stdout of console_style

| Stderr of console_style

val sexp_of_console_output : console_output -> Sexplib0.Sexp.t

val console_output_of_sexp : Sexplib0.Sexp.t -> console_output

val setup_via_params : ?default_output_level:Async.Log.Level.t -> log_to_console_by_default:console_output -> log_to_syslog_by_default:bool -> ?log_to_file_by_default:string -> unit -> unit Async.Command.Param.t

setup_via_params either sets up console, syslog, or file logging with the defaults passed in as parameters to this function, or overrides those defaults via the included command-line parameters.