Module Log_extended.Console

val output : ?debug:Console.Ansi.attr list -> ?info:Console.Ansi.attr list -> ?error:Console.Ansi.attr list -> Async.Writer.t -> Async.Log.Output.t

returns a Log.Output.t given optional styles (i.e. values of type Ansi.t list) for each of the `Debug, `Info, and `Error log levels. The default styling is to display debug messages in yellow, error messages in red, and info messages without any additional styling.

create doesn't take a format argument because colorized output should be read by humans.

module Blocking : sig ... end