Module Reader.Expert

val read_one : [< Output.Format.machine_readable ] -> Async_unix.Reader.t -> Message.t Async_unix.Reader.Read_result.t Async_kernel.Deferred.t

read_one format reader reads a single log message from the reader, advancing the position of the reader to the next log entry.