Module Global_intf.For_testing

val use_test_output : ?map_output:( string -> string ) -> unit -> unit

Change the output of the global log so that it only prints the bodies of messages to stdout, discarding any information about tags, levels, or timestamps. map_output can be used to transform messages before they make it to stdout; by default it is Fn.id.

This is equivalent to: Log.Global.set_output [ Log.For_testing.create_output ~map_output ]